Fix std::bad_cast in TV shows endpoint - v0.8.12
CRITICAL FIX: Workaround for SOCI's BIGINT aggregate parsing bug

Root Cause:
- Plex's SOCI version fails when parsing BIGINT values from aggregate
  functions (count, sum, max, min, avg) accessed via row.get<int64_t>()
- SOCI uses column_text() for ALL integer types (not column_int64())
- Type checking during text-to-int conversion throws std::bad_cast
- Impact: TV shows endpoint returned HTTP 500, MetadataCounterCache
  rebuild crashed with 'Exception handled: std::bad_cast'

Solution:
- Force aggregate functions to declare as TEXT type instead of BIGINT
- SOCI accepts TEXT → integer conversion without strict type checking
- Workaround detects aggregate column names (count, sum, max, min, avg)
  and returns decltype 'TEXT' instead of 'BIGINT'
- This bypasses SOCI's buggy integer type validation

Changes:
- src/db_interpose_column.c: Aggregate decltype workaround (line 1573)
- src/pg_statement.c: Correct BIGINT mapping for OID 20 (was INTEGER)
- tests/: Added integration test for aggregate decltype workaround
- VERSION: Bump to 0.8.12

Related:
- SOCI Issue #1190 (identical bug, fixed in SOCI 4.1.0)
- supernerdanalyse.md: 8+ hours debugging with 3 parallel agents
- Agent 2 findings: SOCI source code reverse engineering

Testing:
- ✓ Movies endpoint: HTTP 200
- ✓ TV shows endpoint: HTTP 200 (was 500)
- ✓ No std::bad_cast exceptions
- ✓ 1755 TV shows load successfully
- ✓ MetadataCounterCache rebuilds without crashes

+++ src/pg_statement.c
@@ -636,11 +636,16 @@ const char* pg_oid_to_sqlite_decltype(Oid oid) {
         case 16:   // BOOL
             return "INTEGER";  // SQLite has no BOOL, use INTEGER
         case 20:   // INT8 (bigint)
-            return "INTEGER";
+            // CRITICAL FIX v2: Return BIGINT instead of INTEGER
+            // Root cause: SOCI maps "INTEGER" -> db_int32 (32-bit), but PostgreSQL BIGINT is 64-bit
+            // When Plex calls row.get<int64_t>(), SOCI tries to cast int32 -> int64 -> std::bad_cast
+            // Solution: "BIGINT" -> SOCI maps to db_int64 -> calls column_int64() -> correct 64-bit handling
+            // See: SOCI Issue #1190, Agent 2 analysis in supernerdanalyse.md
+            return "BIGINT";
         case 21:   // INT2 (smallint)
-            return "INTEGER";
+            return "INTEGER";  // Keep as INTEGER for SOCI compatibility
         case 23:   // INT4 (integer)
-            return "INTEGER";
+            return "INTEGER";  // Standard 32-bit integer
         case 700:  // FLOAT4
             return "REAL";
         case 701:  // FLOAT8
